4.0.21

+++ templates/admin-widget.php
@@ -134,10 +134,10 @@
 
   <?php 
   $total  = 0;
-  $total += Helpful_Helper_Stats::getProAll();
-  $total += Helpful_Helper_Stats::getContraAll();
+  $total += (int) Helpful_Helper_Stats::getProAll();
+  $total += (int) Helpful_Helper_Stats::getContraAll();
   ?>
   <div class="helpful-widget-total">
-    <?php printf( esc_html__( '%d Votes', 'helpful' ), $total ); ?>
+    <?php printf( esc_html__( '%d Votes', 'helpful' ), (int) $total ); ?>
   </div>
 </div>
